% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{alate_infect_sims}
\alias{alate_infect_sims}
\title{Proportion of alates and plants that are inoculated, for a given landscape.}
\usage{
alate_infect_sims(
  x_size,
  y_size,
  delta_a,
  delta_p,
  land,
  alate,
  infected,
  x,
  y
)
}
\arguments{
\item{x_size}{Single integer indicating x dimension of landscape.
Locations must be between \code{1} and \code{x_size}.}

\item{y_size}{Single integer indicating y dimension of landscape.
Locations must be between \code{1} and \code{y_size}.}

\item{delta_a}{Single numeric indicating the probability that an
uninoculated alate is loaded with a virus if it interacts with an
inoculated plant.}

\item{delta_p}{Single numeric indicating the probability that an
uninoculated plant is loaded with a virus if it interacts with an
inoculated alate.}

\item{land}{Integer vector indicating the focal landscape.}

\item{alate}{Integer vector indicating the focal alate.}

\item{infected}{Logical vector indicating whether focal plant is
infected with virus.}

\item{x}{Integer vector indicating location of the focal plant.
All values must be between \code{1} and \code{x_size}.}

\item{y}{Integer vector indicating location of the focal plant.
All values must be between \code{1} and \code{y_size}.}
}
\description{
ASSUMPTIONS OF THIS FUNCTION:
}
\details{
\enumerate{
\item Vectors \code{land}, \code{alate}, \code{infected}, \code{x}, and \code{y} are all sorted first by
\code{land}, then by \code{alate}, then by time (not an argument).
\item On the timeline of these simulations, even if an alate inoculates
a plant with a virus, the virus doesn't have enough time to replicate
inside the plant to cause a subsequent alate to get infected.
\item All landscapes have the same number of alates and landscape dimensions
(\code{x_size} and \code{y_size}).
}
}
