% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{alate_search_sims}
\alias{alate_search_sims}
\title{Simulate aphid alates seeking host plants.}
\usage{
alate_search_sims(
  max_t,
  plant_xy,
  plant_types,
  alpha,
  beta,
  epsilon = 1,
  w = 0.2,
  radius = 7.336451,
  xy0 = NULL,
  randomize_xy0 = TRUE,
  n_alates = 1L,
  summarize = "none",
  show_progress = FALSE,
  n_threads = 1L
)
}
\arguments{
\item{max_t}{Single integer indicating the number of time steps to simulate.}

\item{plant_xy}{Two-column integer matrix containing x and y coordinates
for plants. Note that all locations from 1 to the max in each dimension
must be represented by a plant.}

\item{plant_types}{Integer vector with the same number of items as the
number of rows in \code{plant_xy} indicating the type of plant each
plant is. Values in this vector must be 1 (virus-infected),
2 (\emph{Pseudomonas}-infected), 3 (none), or 4 (both).
Not all values must be present in the landscape.
Note that all locations from 1 to the max in each dimension
must be represented by a plant.}

\item{alpha}{Effect of virus infection on alate alighting.
Values \verb{> 1} cause alates to be attracted to virus-infected plants,
while values \verb{< 1} cause them to be repelled by virus-infected plants.
Values must be \verb{> 0}.}

\item{beta}{Effect of \emph{Pseudomonas} infection on alate alighting.
Values \verb{> 1} cause alates to be attracted to \emph{Pseudomonas}-infected plants,
while values \verb{< 1} cause them to be repelled by \emph{Pseudomonas}-infected plants.
Values must be \verb{> 0}.}

\item{epsilon}{Effect of virus infection on alate acceptance.
Values \verb{> 1} cause alates to be more likely to stay and feed
(indefinitely) on virus-infected plants,
while values \verb{< 1} cause them to be less likely to stay and feed on
virus-infected plants.
Values must be \verb{> 0}, and \code{epsilon * w} must be \verb{< 1}.}

\item{w}{Probability that an alate accepts a plant, meaning that
it stays to feed on it indefinitely.
Must be \verb{> 0} and \verb{< 1}. Defaults to \code{0.2}.}

\item{radius}{Max distance that alates will travel between plants.
Defaults to \code{7.336451}, which is based on previous work.
See "Radius" section below for details.}

\item{xy0}{Numeric vector of length 2 indicating starting x and y
coordinates for alate(s). When \code{xy0 = NULL} and
\code{randomize_xy0 = FALSE}, alate(s) start at the middle of the
search space (x = 0, y = 0).
This function throws an error when \code{xy0} is provided and
\code{randomize_xy0 = TRUE} because these are conflicting.
Defaults to \code{NULL}.}

\item{randomize_xy0}{Single logical for whether to randomize starting
coordinates for alate(s). If \code{TRUE}, x and y coordinates are
randomly chosen from the landscape.
If \code{FALSE}, alates start in the middle of the landscape.
Defaults to \code{TRUE}.}

\item{n_alates}{Single integer indicating the number of independent
alates to simulate.
Defaults to \code{1L}.}

\item{summarize}{Single string for whether and how to summarize output.
Options are \code{"none"}, \code{"types"}, or \code{"xy"}.
See below for details on how this changes the output.
Defaults to \code{"none"}.}

\item{show_progress}{Single logical for whether to show progress bar.
Defaults to \code{FALSE}.}

\item{n_threads}{Single integer for the number of threads to use.
Ignored if \code{n_alates == 1}.
Defaults to \code{1L}.}
}
\value{
If \code{summarize = "none"}, a tibble with the columns
\code{alate} (alate number),
\code{time} (time), \code{x} (x coordinate), \code{y} (y coordinate),
\code{to} (the plant type the alate traveled to), and
\code{from} (the plant type the alate traveled from).
If \code{summarize = "types"}, then it outputs a tibble with the columns
\code{to}, \code{from}, and
\code{n} (the number of times all searchers traveled from this plant type to
this other type).
If \code{summarize = "xy"}, then it outputs a tibble with the columns
\code{alate}, \code{x} (the final x coordinate), and
\code{y} (the final y coordinate).
}
\description{
Simulate aphid alates seeking host plants.
}
\section{Radius}{
From "The Role of Aphid Behaviour in the Epidemiology of Potato Virus Y:
a Simulation Study" by Thomas Nemecek (1993; p. 72), dispersal distances
follow a Weibull distribution with shape = 0.6569 and scale = 9.613.

The default for the \code{radius} argument uses the median of this
distribution.
I'm dividing by 0.75 to convert from meters to plant locations that are
0.75 meters apart (typical spacing for pea):
\code{radius = qweibull(0.5, 0.6569, 9.613) / 0.75}.
}

