% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aphid-wasp-classes.R
\name{all_fields}
\alias{all_fields}
\title{Create AllFields object containing aphid, wasp, and field parameters}
\usage{
all_fields(
  clonal_lines,
  wasps,
  n_fields = 2,
  environ_error = FALSE,
  aphid_demog_error = FALSE,
  wasp_demog_error = FALSE,
  K = 1/0.000467,
  K_y_mult = 1/1.57,
  aphid_density_0 = 32,
  wasp_density_0 = c(3, 0),
  mummy_density_0 = 0,
  wasp_delay = 8,
  wasp_disp_m0 = 0.3,
  wasp_disp_m1 = 0.34906,
  wasp_field_attract = 1,
  pred_rate = 0,
  alate_field_disp_p = 0.1,
  constant_wasps = FALSE,
  extinct_N = 1
)
}
\arguments{
\item{clonal_lines}{A \code{multiAphid} object containing the aphid-line-specific
info for all the lines in the simulations.
Each object in the \code{multiAphid} is an \code{aphid} object that results from
the \code{clonal_line} function.
Combine them using \code{c(aphid_obj1, aphid_obj2)}.}

\item{wasps}{A \code{WaspPop} object containing the parasitoid wasp info
for the simulations.}

\item{n_fields}{The number of fields to simulate.
Both wasps and aphids operate separately across fields but can be
connected via dispersal (see arguments \code{wasp_disp_m0}, \code{wasp_disp_m1},
and \code{alate_field_disp_p}).
Defaults to \code{2}.}

\item{environ_error}{Logical for whether to have environmental stochasticity.
This argument applies to both aphids and wasps.
Defaults to \code{FALSE}.}

\item{aphid_demog_error}{Logical for whether to have demographic
stochasticity for aphids. Defaults to \code{FALSE}.}

\item{wasp_demog_error}{Logical for whether to have demographic
stochasticity for wasps. Defaults to \code{FALSE}.}

\item{K}{Aphid density dependence. Either a single number for all fields,
or a vector of length \code{n_fields} to define separate values by field.
Defaults to \code{12.5e3} because this caused simulations to
approximately match experiments.}

\item{K_y_mult}{The number multiplied by \code{K} to get density dependence for
parasitized aphids.
Either a single number for all fields, or a vector of length \code{n_fields}
to define separate values by field.
Defaults to \code{1 / 1.57}, which is from Meisner et al. (2014).}

\item{aphid_density_0}{The total abundance of aphids in each field.
Either a single number for all lines in all fields, or a
\code{n_fields}-row matrix with a column for each aphid line to define
separate values by line and field.
Defaults to \code{32}.}

\item{wasp_density_0}{The abundance of adult wasps in each field.
Either a single number for all fields, or a vector of length \code{n_fields}
to define separate values by field.
Defaults to \code{c(0, 3)}.}

\item{mummy_density_0}{Either a single number for all fields, or a vector of length \code{n_fields}
to define separate values by field.
Defaults to \code{0}.}

\item{wasp_delay}{Number of days to wait to add wasps. This can be useful
to prevent transient dynamics from causing early extinction of
aphid clone(s).
Either a single number for all fields, or a vector of length \code{n_fields}
to define separate values by field.
Defaults to \code{8}.}

\item{wasp_disp_m0}{Proportion of adult wasps from each field that
are added to the dispersal pool when there are no aphids present.
After adding wasps to the pool, they are then evenly distributed
to all fields unless values for \code{wasp_field_attract} are provided.
Defaults to \code{0.3}.}

\item{wasp_disp_m1}{Effect of aphid density on wasp emigration from a patch.
Emigration is \code{wasp_disp_m0 * exp(-wasp_disp_m1 * log(z))}, where \code{z} is
the total number of living aphids in the patch.
Note that if \code{wasp_disp_m0 = 0}, then this parameter doesn't change
anything.
Defaults to \code{0.34906}.}

\item{wasp_field_attract}{Relative attractiveness of fields to wasps.
This affects the proportion of wasps that immigrate from the dispersal
pool to each field.
It doesn't change the number of wasps that leave fields.
This can be a single numeric or a numeric vector of length \code{n_fields}.
If a single numeric is provided, all fields are equally attractive
to wasps.
If a vector is provided, then the vector is divided by its sum
(to make it sum to 1), then those values are used as the proportion of
wasps from the dispersal pool that immigrate to each field.
Note that if \code{wasp_disp_m0 = 0}, then this parameter doesn't change
anything.
Defaults to \code{1}.}

\item{pred_rate}{Daily predation rate on aphids and mummies.
Defaults to \code{0.1}.}

\item{alate_field_disp_p}{Proportion of alates from each field that
are added to the dispersal pool.
After adding alates to the pool, they are then evenly distributed
to all fields.
Defaults to \code{0.1}.}

\item{constant_wasps}{Logical for whether to keep adult wasps at the same
density throughout simulations.
This can be a single logical or a \code{n_fields}-length vector.
Defaults to \code{FALSE}.}

\item{extinct_N}{Threshold below which a line is considered extinct.
Defaults to \code{1}.}
}
\value{
List containing an external pointer of a C++ object that can be
passed onto \code{sim_fields}.
}
\description{
Create AllFields object containing aphid, wasp, and field parameters
}
\examples{
wp <- wasp_pop()
sa <- clonal_line("susceptible")
ra <- clonal_line("resistant", resistant = TRUE, repro_apterous = "low")
af <- all_fields(c(sa, ra), wp)


}
