% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{carrying_capacity}
\alias{carrying_capacity}
\title{Calculate carrying capacity from aphid and patch info}
\usage{
carrying_capacity(
  apterous,
  alates,
  alate_prop,
  disp_prop,
  disp_mort,
  disp_start,
  K
)
}
\arguments{
\item{apterous}{Leslie matrix for apterous aphids of this line}

\item{alates}{Leslie matrix for alates of this line}

\item{alate_prop}{The proportion of new aphids that are alates}

\item{disp_prop}{The proportion of alates that disperse away from the patch.}

\item{disp_mort}{Mortality rate for dispersing alates.}

\item{disp_start}{Index for the age at which alates disperse (0-indexed).}

\item{K}{The "carrying capacity" for this patch.}
}
\description{
Calculate carrying capacity from aphid and patch info
}
