% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{clonal_line}
\alias{clonal_line}
\title{Organize clonal line information}
\usage{
clonal_line(
  name,
  density_0,
  resistant = FALSE,
  surv_juv_apterous = NULL,
  surv_adult_apterous = NULL,
  repro_apterous = NULL,
  surv_juv_alates = "low",
  surv_adult_alates = "low",
  repro_alates = "low",
  surv_paras = "low",
  temp = "low",
  p_instar_smooth = 0.5
)
}
\arguments{
\item{name}{String for clonal line name.}

\item{density_0}{A single number or 2-column matrix indicating starting
aphid densities.
If a number, there will be no starting alates and a total apterous
density equal to the number provided.
Stages will be calculated based on the stable age distribution.
If a matrix, it must have 5 rows (rows indicate instar)
or a row per aphid age (rows indicate age in days).
Matrix column indicates apterous vs alate.}

\item{resistant}{Logical or length-2 vector of survivals of
singly attacked and multiply attacked aphids.
If a logical, \code{FALSE} is equivalent to \code{c(0,0)} and results in no
resistance.
\code{TRUE} results in the resistance values for a resistance line
from unpublished work by Anthony Ives.
Defaults to \code{FALSE}.}

\item{surv_juv_apterous}{A single number for the juvenile survival rate for
apterous aphids. Defaults to \code{NULL}, which results in estimates
from a medium-reproduction line.}

\item{surv_adult_apterous}{A vector of adult survival probabilities for
apterous aphids. Defaults to \code{NULL}, which results in estimates
from a medium-reproduction line.}

\item{repro_apterous}{A vector of fecundities for for apterous aphids.
Defaults to \code{NULL}, which results in estimates from a
medium-reproduction line.}

\item{surv_juv_alates}{A single number for the juvenile survival rate for
alates aphids. Defaults to \code{"low"}, which results in estimates
from a low-reproduction line.}

\item{surv_adult_alates}{A vector of adult survival probabilities for
alates aphids. Defaults to \code{"low"}, which results in estimates
from a low-reproduction line.}

\item{repro_alates}{A vector of fecundities for for alates aphids.
Defaults to \code{"low"}, which results in estimates from a low-reproduction
line.}

\item{surv_paras}{A single number for the juvenile survival rate for
parasitized aphids.
Because parasitized aphids don't make it to adulthood, this is the only
survival rate necessary.
Defaults to \code{"low"}, which results in estimates from a
low-reproduction line.}

\item{temp}{Single string specifying \code{"low"} (20º C) or \code{"high"} (27º C)
temperature. Defaults to \code{"low"}.}

\item{p_instar_smooth}{A value greater than zero here makes not all 4th
instar aphids go to the adult stage at the same time.
Of the 4th instars that would've moved to adulthood, \code{p_instar_smooth}
remain as 4th instars.
To keep the growth rate approximately equal, for the aphids that
would've transitioned to 4th instar, \code{p_instar_smooth} move to
adulthood instead.
Defaults to \code{0.5}.}
}
\value{
A list with the necessary info to pass onto sim_gameofclones.
}
\description{
To smooth stage structure out over time,
not all 4th instar aphids move to adulthood immediately.
If adulthood starts on day \code{t}, then half of aphids at age \code{t-2} move to
adulthood, and half at age \code{t-1} do, too.
I adjusted age \code{t-2}, too, to avoid this affecting the growth rate too much.
}
