% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clonewars-package.R
\docType{package}
\name{clonewars-package}
\alias{clonewars-package}
\alias{clonewars}
\title{The 'clonewars' package.}
\description{
Analyses and simulations related to repeated aphid clonal
evolution experiments.
}
