% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{compare_priors}
\alias{compare_priors}
\title{Visually compare normal, student_t, cauchy, laplace, and product_normal.}
\usage{
compare_priors(location = 0, scale = 1, df_t = 2, xlim = c(-10, 10))
}
\arguments{
\item{scale}{Scale for all}

\item{df_t}{Degrees of freedom for `student_t`.}

\item{xlim}{X limit for the plot.}
}
\description{
From \url{http://mc-stan.org/rstanarm/reference/priors.html#examples}
}
