% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_models.R
\name{fit_one_ts}
\alias{fit_one_ts}
\title{Fit one aphid-growth time series.}
\usage{
fit_one_ts(X, data = sys.frame(sys.parent()), ...)
}
\arguments{
\item{X}{Name of the object in `data` that contains the log-transformed counts
through time. `NA`s are silently removed.}

\item{data}{An optional data frame, list, or environment that contains `X`.
By default, variables are taken from the environment from which
the function was called.}

\item{...}{Arguments passed to `rstan::sampling` (e.g., iter, chains).}
}
\value{
A `stanfit` object containing the model fit.
}
\description{
Fit one aphid-growth time series.
}
\examples{

R0 <- 0.25
alpha <- 0.002
sigma_process <- 0.1

X <- as.numeric(sim_lines(R0, alpha, n_reps = 1, nobs_ts = 20, process = 0.1))

# (You would normally not change iter to 100)
fit <- fit_one_ts(X, iter = 100)


}
