% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudogameofclones-package.R
\docType{package}
\name{pseudogameofclones-package}
\alias{pseudogameofclones-package}
\alias{pseudogameofclones}
\title{The 'pseudogameofclones' package.}
\description{
Analyses and simulations related to aphid--parasitoid
eco-evolutionary dynamics experiments.
}
