% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{leslie}
\alias{leslie}
\title{Leslie matrices for fast and slow aphid populations.}
\format{A list with two 27 x 27 matrices.}
\source{
\url{https://onlinelibrary.wiley.com/doi/abs/10.1890/13-1933.1}
}
\usage{
leslie
}
\description{
Leslie matrices for fast and slow aphid populations.
}
\keyword{datasets}
