% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manip.R
\name{line_data}
\alias{line_data}
\title{Convert a data frame of data to one to be used in `fit_lines`.}
\usage{
line_data(data, line, rep, date, X)
}
\arguments{
\item{data}{A data frame.}

\item{line}{Name of column (no quotes) indicating the aphid line.}

\item{rep}{Name of column (no quotes) indicating the rep within each line.}

\item{date}{Name of column (no quotes) indicating the date.}

\item{X}{Name of column (no quotes) indicating the log(N).}
}
\value{
A list containing (1) a matrix of log(N), where each time series has its own
    column (named `X` in the output list) and (2) a vector of integers indicating
    the aphid-line number for each time-series column in (named `line_ts`).
    Aphid line numbers come from converting the original factor column into
    an integer. You can replicate the same integers by running `as.integer(line)`.
}
\description{
Convert a data frame of data to one to be used in `fit_lines`.
}
