% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manip.R
\name{load_prior_data}
\alias{load_prior_data}
\title{Load data that won't be used for the actual analysis, to develop priors.}
\usage{
load_prior_data(filter_pars = NULL, file = NA)
}
\arguments{
\item{filter_pars}{A list with the names \code{"start"} and \code{"end"}, containing single
numbers with threshold for filtering the beginning and ending of time series,
respectively. Set to \code{NULL} to avoid filtering entirely.
Defaults to \code{list(start = 0.0, end = 0.9)}.}

\item{file}{A filename to read from. If left empty, this will read from the
default path on your DropBox folder.}
}
\description{
Argument \code{filter_pars} defaults to the same as for \code{load_data}.
}
