% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manip.R
\name{load_pz_data}
\alias{load_pz_data}
\title{Load data from PZ counts that were almost finished.}
\usage{
load_pz_data(filter_pars = list(start = 0, end = 0.8),
  remove_unfinished = TRUE, allow_NA = TRUE, impute_fxn = NULL,
  file = NA)
}
\arguments{
\item{filter_pars}{A list with the names \code{"start"} and \code{"end"}, containing single
numbers with threshold for filtering the beginning and ending of time series,
respectively. Set to \code{NULL} to avoid filtering entirely.
Defaults to \code{list(start = 0.5, end = 1.0)}.}

\item{allow_NA}{Boolean for whether time series with NAs should be included.
Defaults to \code{TRUE}.}

\item{file}{A filename to read from. If left empty, this will read from the
default path on your DropBox folder.}
}
\description{
Here, I'm adding a last date that has 0.75 of the max.
This function is temporary and should be removed when counts are finished.
}
