% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{logit}
\alias{logit}
\alias{inv_logit}
\title{Logit and inverse logit functions.}
\usage{
logit(p)

inv_logit(a)
}
\description{
Logit and inverse logit functions.
}
\section{Functions}{
\itemize{
\item \code{inv_logit}: 
}}

