% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manip.R
\name{make_pred_df}
\alias{make_pred_df}
\title{Make prediction data frame from model output and the original data frame.}
\usage{
make_pred_df(stan_fit, orig_data, line, rep, alpha = 0.05)
}
\arguments{
\item{stan_fit}{The \code{stanfit} object containing the model fit.}

\item{orig_data}{The original data frame.}

\item{line}{An optional parameter specifying the name of the column in \code{orig_data}
that contains info on the aphid line.}

\item{rep}{An optional parameter specifying the name of the column in \code{orig_data}
that contains info on the rep within each aphid line.}
}
\description{
Make prediction data frame from model output and the original data frame.
}
