% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{plant_death}
\alias{plant_death}
\title{Information for simulating plant death.}
\format{A list with two fields, each containing a single number:
\itemize{
\item \code{days_to_max}: An integer specifying the number of days after which the plant
begins to die.
\item \code{after_max_mortality}: The per-day mortality to add to the aphid populations on
a plant after the plant begins to die.
}}
\usage{
plant_death
}
\description{
Information for simulating plant death.
}
\keyword{datasets}
