% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudogameofclones-package.R
\docType{package}
\name{pseudogameofclones-package}
\alias{pseudogameofclones}
\alias{pseudogameofclones-package}
\title{pseudogameofclones: Clonal Evolution in Aphid Metapopulations}
\description{
Simulate pea aphid clonal evolution and parasitoids through time and space.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/lucasnell/pseudogameofclones}
  \item Report bugs at \url{https://github.com/lucasnell/pseudogameofclones/issues}
}

}
\author{
\strong{Maintainer}: Lucas A. Nell \email{lucnell@gmail.com} (\href{https://orcid.org/0000-0003-3209-0517}{ORCID}) [copyright holder]

}
\keyword{internal}
