% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rel_res_fitness}
\alias{rel_res_fitness}
\title{Calculate r_r / r_s (relative fitness for resistance vs susceptible aphids).}
\usage{
rel_res_fitness(para, p_res = 0.48)
}
\arguments{
\item{para}{Numeric vector of the proportion of aphids that are parasitized.}

\item{p_res}{The assumed proportion of aphids that are resistant to
parasitism. This is the mean proportion from Ives et al. (2020).}
}
\value{
A numeric vector of r_r / r_s (relative fitness for resistance vs
susceptible aphids)
}
\description{
This is almost directly from code in Ives et al. (2020).
}
\details{
This function will be very slow if used inside an \code{apply} function
because it already uses \code{sapply} within.
Instead, \code{para} should be a vector.
}
