% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{restart_experiment}
\alias{restart_experiment}
\title{Restart experimental simulations.}
\usage{
restart_experiment(
  sims_obj,
  new_starts = NULL,
  stage_ts_out = FALSE,
  max_t = 250,
  save_every = 1,
  alate_field_disp_p = NULL,
  K = NULL,
  alate_b0 = NULL,
  alate_b1 = NULL,
  K_y_mult = NULL,
  s_y = NULL,
  a = NULL,
  k = NULL,
  h = NULL,
  wasp_disp_m0 = NULL,
  wasp_disp_m1 = NULL,
  wasp_field_attract = NULL,
  mum_smooth = NULL,
  pred_rate = NULL,
  sep_adults = NULL,
  show_progress = NULL,
  perturb = NULL,
  no_warns = FALSE,
  n_threads = 1L
)
}
\arguments{
\item{sims_obj}{A \code{cloneSims} object output from \code{sim_experiments}.}

\item{new_starts}{A dataframe or list of dataframes indicating the
new starting abundances for all populations (wasps, mummies,
all aphid lines) and stages.
It should be the exact same format as what's in \code{sims_obj$all_info}.
(But don't change \code{sims_obj$all_info} to make this be true!)}

\item{stage_ts_out}{Single logical for whether to output stage-structured
information for all time points.
If \code{TRUE}, the output object will contain this information in the
\code{stage_ts} field, which will be a list of data frames.
Defaults to \code{FALSE}.}
}
\description{
Note for most of this function's arguments, their default value is \code{NULL},
which results in them being the same as for the original simulations.
The exceptions to this are the arguments \code{sims_obj}, \code{new_starts},
\code{stage_ts_out}, \code{max_t}, \code{save_every}, \code{perturb}, and \code{no_warns}.
Arguments \code{sims_obj}, \code{new_starts}, and \code{stage_ts_out} are new to this
function (see their documentation below for details).
Arguments \code{max_t} and \code{save_every} are set to \code{250} and \code{1}, respectively,
which results in a shorter time series with greater resolution, the typical
use case for this function.
Argument \code{perturb} has a default value of \code{NULL} which results in no
perturbations.
You can input a new data frame if you want to perturb these restarted
experiments.
If the initial simulations had perturbations but the new one won't,
this function will provide a warning indicating this.
You can suppress this warning by setting \code{no_warns = TRUE}.
}
