% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rm_tibs}
\alias{rm_tibs}
\title{Remove tibbles from a cloneSims or cloneSimsRestart object}
\usage{
rm_tibs(sims_obj)
}
\arguments{
\item{sims_obj}{A cloneSims or cloneSimsRestart object}
}
\value{
The same class of object as input, but with the \code{aphids} and \code{wasps} fields
converted to data frames instead of tibbles.
}
\description{
Remove tibbles from a cloneSims or cloneSimsRestart object
}
