% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sad_leslie}
\alias{sad_leslie}
\title{Compute the "stable age distribution" from a Leslie matrix}
\usage{
sad_leslie(leslie)
}
\arguments{
\item{leslie}{Leslie matrix of the population.}
}
\value{
Vector with the stable age distribution given the Leslie matrix.
}
\description{
Compute the "stable age distribution" from a Leslie matrix
}
