% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_models.R
\name{sim_cages}
\alias{sim_cages}
\title{Simulate cages with the same starting conditions.}
\usage{
sim_cages(X_0, max_t, R, A, D_slope, D_inter, process_error, n_reps,
  n_chains = 1, ...)
}
\arguments{
\item{X_0}{Matrix of \code{log(# aphids)} at time t=0. Should be matrix of size
\code{n_plants} x \code{n_lines}.}

\item{max_t}{Time steps per cage. Single integer.}

\item{R}{Max growth rates per aphid line. Vector of length \code{n_lines}.}

\item{A}{Density dependence per aphid line. Vector of length \code{n_lines}.}

\item{D_slope}{Dispersal slope per aphid line. Vector of length \code{n_lines}.}

\item{D_inter}{Dispersal intercept per aphid line. Vector of length \code{n_lines}.}

\item{process_error}{SD of process error. Single double.}

\item{n_reps}{Number of reps (i.e., cages) per chain.}

\item{n_chains}{Number of chains to use. Note that values >1 will result in
extra info being printed to the console. You can ignore this. Defaults to \code{1}.}

\item{...}{Other options passed to \code{rstan::sampling()}.}
}
\description{
Simulate cages with the same starting conditions.
}
