% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{sim_clonewars}
\alias{sim_clonewars}
\title{Simulate multiple reps and simplify output.}
\usage{
sim_clonewars(
  n_reps,
  clonal_lines,
  n_patches = 4,
  max_t = 100,
  plant_check_gaps = c(3, 4),
  max_plant_age = 1e+06,
  clear_surv = 0,
  max_N = 0,
  temp = "low",
  no_error = FALSE,
  disp_error = FALSE,
  environ_error = FALSE,
  plant_K_error = FALSE,
  wither_effects_error = FALSE,
  sigma_x = environ$sigma_x,
  sigma_y = environ$sigma_y,
  mean_K = 1806.176,
  sd_K = 658.4736,
  K_y_mult = 1/1.57,
  death_prop = 0.8,
  rho = environ$rho,
  a = wasp_attack$a,
  k = wasp_attack$k,
  h = wasp_attack$h,
  wasp_density_0 = 4,
  wasp_delay = 0,
  sex_ratio = populations$sex_ratio,
  s_y = populations$s_y,
  rel_attack = NULL,
  mum_density_0 = 0,
  pred_rate = 0,
  disp_rate = 1,
  disp_mort = 0,
  alate_b0 = -2.988,
  alate_b1 = 0,
  shape1_death_mort = 3.736386,
  shape2_death_mort = 5.777129,
  extinct_N = 1,
  save_every = 1,
  n_threads = max(parallel::detectCores() - 2, 1),
  show_progress = FALSE
)
}
\arguments{
\item{n_reps}{Number of reps to simulate.}

\item{n_patches}{Number of patches to simulate.}

\item{max_t}{Max time points to simulate for each rep.}

\item{disp_error}{Boolean for whether to include dispersal stochasticity.}

\item{extinct_N}{Threshold below which a line is considered extinct.}

\item{save_every}{Abundances will be stored every \code{save_every} time points.}

\item{show_progress}{Boolean for whether to show progress bar. Defaults to
\code{FALSE}.}

\item{N_0}{Starting abundances for each aphid line on each patch.
Can be a single number if you want the same value for each line on each patch,
or a matrix if you want to specify everything.}

\item{R}{Growth rates for each line.}

\item{A}{Density dependence for each line.}

\item{D_vec}{Vector of \code{b0} values, where the predicted number of dispersed aphids is
given by \code{exp(b0) * N} when \code{N} is the number of total aphids.}

\item{process_error}{SD of process error. Set to 0 for no process error.}

\item{log_zeta_mean}{Mean of the distribution of log(zeta) values.}

\item{log_zeta_sd}{SD of the distribution of log(zeta) values.}

\item{zeta_t_thresh}{Threshold for \code{exp(zeta * (t - mu_time))} that makes that
patch get replaced. This is equivalent to the threshold for patch "health"
that would make an experimenter replace it.}

\item{mu_time}{Mean of time values.}

\item{repl_times}{Vector of times at which to replace patches.}

\item{repl_threshold}{Threshold above which patches are replaced.}

\item{by_patch}{Logical for whether to summarize abundances by patch, rather
than separately by line and patch.}

\item{n_cores}{Number of cores to use. Defaults to \code{1}.}

\item{line_names}{Vector of names to assign to lines.}
}
\description{
Simulate multiple reps and simplify output.
}
