% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{sim_clonewars_full}
\alias{sim_clonewars_full}
\title{Simulate multiple reps and simplify output - all options.}
\usage{
sim_clonewars_full(
  n_reps,
  clonal_lines,
  n_fields = 2,
  n_plants = 1,
  max_t = 250,
  plant_check_gaps = 1,
  max_plant_age = 1e+09,
  clear_surv = 0,
  max_N = 0,
  temp = "low",
  no_error = TRUE,
  disp_error = FALSE,
  environ_error = FALSE,
  plant_K_error = FALSE,
  wilted_effects_error = FALSE,
  sigma_x = environ$sigma_x,
  sigma_y = environ$sigma_y,
  mean_K = 1/0.000467,
  sd_K = 0,
  K_y_mult = 1/1.57,
  wilted_prop = 1.1,
  rho = environ$rho,
  a = wasp_attack$a,
  k = wasp_attack$k,
  h = wasp_attack$h,
  wasp_density_0 = c(3, 0),
  wasp_delay = 8,
  wasp_disp_p = 0,
  sex_ratio = populations$sex_ratio,
  s_y = populations$s_y,
  constant_wasps = FALSE,
  rel_attack = wasp_attack$rel_attack,
  mum_density_0 = 0,
  mum_smooth = 0.4,
  max_mum_density = 0,
  pred_rate = 0,
  alate_plant_disp_p = 0.1,
  disp_mort = 0,
  alate_b0 = logit(0.093),
  alate_b1 = 0,
  alate_field_disp_p = 0.1,
  shape1_wilted_mort = 3.736386,
  shape2_wilted_mort = 5.777129,
  extinct_N = 1,
  save_every = 1,
  n_threads = 1,
  show_progress = FALSE,
  perturb = NULL
)
}
\arguments{
\item{n_reps}{Number of reps to simulate.}

\item{clonal_lines}{A \code{multiAphid} object containing the aphid-line-specific
info for all the lines in the simulations.
Each object in the \code{multiAphid} is an \code{aphid} object that results from
the \code{clonal_line} function.
Combine them using \code{c(aphid_obj1, aphid_obj2)}.}

\item{n_fields}{The number of fields to simulate.
Wasps operate at the field level, whereas aphids operate at the
plant level.
Both wasps and aphids operate separately across fields but can be
connected via dispersal (see arguments \code{wasp_disp_p} and
\code{alate_field_disp_p}).
Defaults to \code{2}.}

\item{n_plants}{The number of plants per field.
This argument is only useful if you want to simulate the process
of plants or groups of plants dying.
Wasps operate at the field level, whereas aphids operate at the
plant level.
"Plant" can indicate a group of plants where the aphids can freely
disperse across them.
Aphids can disperse across plants via alate production and the
`` argument.
Defaults to \code{1}.}

\item{max_t}{How many days to simulate. Defaults to \code{250}.}

\item{plant_check_gaps}{Gap(s) between when plants are check on.
This is used if you want to see what will happen if you can't check
on things every day in an experiment.
You could use \code{c(3, 4)} for checking twice per week, for example.
Note that this argument determines how often
(1) wasps and alates disperse across fields and
(2) plants are check for exceeding \code{max_plant_age} or \code{max_N}.
So using this argument to simulate harvesting fields will also cause
wasp and alate dispersal to not occur daily.
Defaults to \code{1}.}

\item{max_plant_age}{Age at which plants are cleared.
This can be useful to simulate harvesting fields.
A value of \code{0} turns this off completely, but it also requires that
\code{max_N > 0} because the internal code wants some threshold to look for.
Defaults to \code{1e9} which effectively turns this off.}

\item{clear_surv}{Survival of aphids and mummies when a harvest occurs.
Defaults to \code{0}.}

\item{max_N}{Maximum number of aphids at which plants are cleared.
This is useful for simulating an experiment where you have a threshold
at which you harvest plants.
Defaults to \code{0}, which turns this off.}

\item{temp}{A string indicating which temperature to simulate.
Options are \code{"low"} (20ºC) or \code{"high"} (27ºC).
Defaults to \code{"low"}.}

\item{no_error}{Logical for whether to have no error at all.
This being \code{TRUE} overrides all other \verb{_error} arguments.
Defaults to \code{TRUE}.}

\item{disp_error}{Logical for whether to have stochasticity in the
dispersal process.
Defaults to \code{FALSE}.}

\item{environ_error}{Logical for whether to have environmental stochasticity
Defaults to \code{FALSE}.}

\item{plant_K_error}{Logical for whether to have stochasticity in aphid
density dependence across plants.
Defaults to \code{FALSE}.}

\item{wilted_effects_error}{Logical for whether to have stochasticity in
the effects of plants dying on aphid populations.
Defaults to \code{FALSE}.}

\item{sigma_x}{Standard deviation of environmental stochasticity for aphids.
Defaults to the internal object \code{environ$sigma_x},
which is from Meisner et al. (2014).}

\item{sigma_y}{Standard deviation of environmental stochasticity for wasps.
Defaults to the internal object \code{environ$sigma_y},
which is from Meisner et al. (2014).}

\item{mean_K}{Mean of the distribution of \code{K} values (affecting aphid
density dependence) among plants.
Defaults to \code{1 / 4.67e-4}, which is from Meisner et al. (2014).}

\item{sd_K}{Mean of the distribution of \code{K} values (affecting aphid
density dependence) among plants. Defaults to \code{0}.}

\item{K_y_mult}{The number multiplied by \code{K} to get density dependence for
parasitized aphids.
Defaults to \code{1 / 1.57}, which is from Meisner et al. (2014).}

\item{wilted_prop}{Proportion of carrying capacity that causes the plant
to become wilted. Values > 1 cause this to be ignored.
Defaults to \code{1.1}.}

\item{rho}{Environmental correlation among instars.
Defaults to \code{environ$rho}.}

\item{a}{Parasitoid attack rate. Defaults to the internal object
\code{wasp_attack$a}, which is from Meisner et al. (2014).}

\item{k}{Aggregation parameter of the negative binomial distribution.
Defaults to the internal object \code{wasp_attack$k},
which is from Meisner et al. (2014).}

\item{h}{Parasitoid handling time. Defaults to the internal object
\code{wasp_attack$h}, which is from Meisner et al. (2014).}

\item{wasp_density_0}{Starting adult wasp density.
Must be a single number or a \code{n_fields}-length vector.
Defaults to \code{c(3, 0)}.}

\item{wasp_delay}{Delay in days between when the aphids start and the
wasps are added.
This can be a single integer or a \code{n_fields}-length vector.
Defaults to \code{8}.}

\item{wasp_disp_p}{Proportion of adult wasps from each field that
are added to the dispersal pool.
After adding wasps to the pool, they are then evenly distributed
to all fields.
This happens only on days indicated by \code{plant_check_gaps}.
Defaults to \code{0}.}

\item{sex_ratio}{Sex ratio of adult wasps. Defaults to \code{0.5}.}

\item{s_y}{Daily survival rate of adult wasps.
Defaults to \code{populations$s_y}, which is from Meisner et al. (2014).}

\item{constant_wasps}{Logical for whether to keep adult wasps at the same
density throughout simulations.
This can be a single logical or a \code{n_fields}-length vector.
Defaults to \code{FALSE}.}

\item{rel_attack}{Relative parasitoid attack rate among instars.
Defaults to \code{wasp_attack$rel_attack}, which is from
Meisner et al. (2014).}

\item{mum_density_0}{Starting mummy density. Defaults to \code{0}.}

\item{mum_smooth}{Proportion of mummies that will NOT take exactly 3 days
to develop. As this value approaches 2/3, it will provide greater
smoothing of wasp numbers through time.
Defaults to \code{0.4}.}

\item{max_mum_density}{Maximum mummy density (ignored if zero).
Used to test the effects of a potential experimental treatment.
Defaults to \code{0}.}

\item{pred_rate}{Daily predation rate on aphids and mummies. Defaults to \code{0}.}

\item{alate_plant_disp_p}{Proportion of alates that emigration to new plants
every day. Note that this does not depend on \code{plant_check_gaps}
like among-field alate dispersal does.
This is because we would have to replicate among-field dispersal
by hand in the experiments we were planning, and we wanted to be sure
that not checking daily wouldn't affect anything significantly.
It didn't.
This can be a single number or a vector with the same length as the
number of aphid lines.
Defaults to \code{0.1}.}

\item{disp_mort}{Mortality rate of alates that disperse among plants.
Defaults to \code{0}.}

\item{alate_b0}{The proportion of offspring from apterous aphids is
\code{inv_logit(alate_b0 + alate_b1 * N)} where \code{N} is the total number of
aphids on that plant.
Defaults to \code{logit(0.3)}.}

\item{alate_b1}{The proportion of offspring from apterous aphids is
\code{inv_logit(alate_b0 + alate_b1 * N)} where \code{N} is the total number of
aphids on that plant.
Defaults to \code{0}, which makes alate production not density dependent.}

\item{alate_field_disp_p}{Proportion of alates from each field that
are added to the dispersal pool.
After adding alates to the pool, they are then evenly distributed
to all fields.
This happens only on days indicated by \code{plant_check_gaps}.
Defaults to \code{0.1}.}

\item{shape1_wilted_mort}{Shape 1 for the beta distribution that generates
mortality parameters for aphid populations living on a wilted plant.
If \code{wilted_effects_error} is \code{FALSE}, then the wilted-plant-induced
mortality is fixed at
\code{shape1_wilted_mort / (shape1_wilted_mort + shape2_wilted_mort)}.
Defaults to \code{3.736386}, which is based on previous work in the lab.}

\item{shape2_wilted_mort}{Shape 2 for the beta distribution that generates
mortality parameters for aphid populations living on a wilted plant.
Defaults to \code{5.777129}, which is based on previous work in the lab.}

\item{extinct_N}{Threshold below which a line is considered extinct.
Defaults to \code{1}.}

\item{save_every}{Abundances will be stored every \code{save_every} time points.
Defaults to \code{1}.}

\item{n_threads}{Number of threads to use if OpenMP is enabled
(ignored otherwise).
Find out whether it's enabled using \code{clonewars:::using_openmp()}.
Defaults to \code{1}.}

\item{show_progress}{Boolean for whether to show progress bar.
Defaults to \code{FALSE}.}

\item{perturb}{Information for perturbing populations in the simulations.
It should be a dataframe with 4 columns:
* \code{when}: Integers indicating at what timepoint(s) to do the
perturbations. These can be repeated if you want to perturb
multiple things at the same time.
* \code{where}: What field to do the perturbations in.
* \code{who}: Which population to perturb.
This can be a character vector where values must be the name of an
aphid line, \code{"wasps"}, or \code{"mummies"}.
It can also be an integer vector where, for \code{n} aphid lines,
values \verb{<= n} indicate an aphid line,
values \verb{== n+1} indicate mummies,
and values \verb{== n+2} indicate adult wasps.
Note that perturbing the mummy population also perturbs the
still-living but parasitized aphids, too.
* \code{how}: Numbers \verb{>= 0} that are multiplied by the desired population
to cause the perturbation.}
}
\description{
This mainly differs from \code{sim_clonewars} in that it allows for
stochasticity and for simulating the process of plants dying and
being replaced.
The latter process doesn't appear important except for very small scales
and for defenseless plants (like fava bean).
Stochasticity isn't necessarily important, either, and the code has
change in other ways such that I have largely abandoned the
stochastic simulations.
So use this at your own risk.
}
