% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{sim_experiments}
\alias{sim_experiments}
\title{Deterministic simulations of multiple fields of aphids and wasps.}
\usage{
sim_experiments(
  clonal_lines,
  n_fields = 2,
  max_t = 250,
  K = 12500,
  alate_b0 = -5,
  alate_b1 = 0.0022,
  alate_field_disp_p = 0.1,
  K_y_mult = 1/1.57,
  s_y = populations$s_y,
  a = wasp_attack$a,
  k = wasp_attack$k,
  h = wasp_attack$h,
  rel_attack = wasp_attack$rel_attack,
  wasp_density_0 = c(3, 0),
  wasp_delay = 8,
  wasp_disp_m0 = 0,
  wasp_disp_m1 = 0,
  wasp_field_attract = 1,
  constant_wasps = FALSE,
  mum_smooth = 0.4,
  pred_rate = 0.1,
  extinct_N = 1,
  save_every = 1,
  perturb = NULL,
  show_progress = FALSE
)
}
\arguments{
\item{clonal_lines}{A \code{multiAphid} object containing the aphid-line-specific
info for all the lines in the simulations.
Each object in the \code{multiAphid} is an \code{aphid} object that results from
the \code{clonal_line} function.
Combine them using \code{c(aphid_obj1, aphid_obj2)}.}

\item{n_fields}{The number of fields to simulate.
Wasps operate at the field level, whereas aphids operate at the
plant level.
Both wasps and aphids operate separately across fields but can be
connected via dispersal (see arguments \code{wasp_disp_m0}, \code{wasp_disp_m1},
and \code{alate_field_disp_p}).
Defaults to \code{2}.}

\item{max_t}{How many days to simulate. Defaults to \code{250}.}

\item{K}{Aphid density dependence.
Defaults to \code{12.5e3} because this caused simulations to
approximately match experiments.}

\item{alate_b0}{The proportion of offspring from apterous aphids is
\code{inv_logit(alate_b0 + alate_b1 * N)} where \code{N} is the total number of
aphids on that plant.
Defaults to \code{-5}.}

\item{alate_b1}{The proportion of offspring from apterous aphids is
\code{inv_logit(alate_b0 + alate_b1 * N)} where \code{N} is the total number of
aphids on that plant.
Defaults to \code{0.0022}, which makes alate production only mildly
density dependent.}

\item{alate_field_disp_p}{Proportion of alates from each field that
are added to the dispersal pool.
After adding alates to the pool, they are then evenly distributed
to all fields.
This happens only on days indicated by \code{plant_check_gaps}.
Defaults to \code{0.1}.}

\item{K_y_mult}{The number multiplied by \code{K} to get density dependence for
parasitized aphids.
Defaults to \code{1 / 1.57}, which is from Meisner et al. (2014).}

\item{s_y}{Daily survival rate of adult wasps.
Defaults to \code{populations$s_y}, which is from Meisner et al. (2014).}

\item{a}{Parasitoid attack rate. Defaults to the internal object
\code{wasp_attack$a}, which is from Meisner et al. (2014).}

\item{k}{Aggregation parameter of the negative binomial distribution.
Defaults to the internal object \code{wasp_attack$k},
which is from Meisner et al. (2014).}

\item{h}{Parasitoid handling time. Defaults to the internal object
\code{wasp_attack$h}, which is from Meisner et al. (2014).}

\item{rel_attack}{Relative parasitoid attack rate among instars.
Defaults to \code{wasp_attack$rel_attack}, which is from
Meisner et al. (2014).}

\item{wasp_density_0}{Starting adult wasp density.
Must be a single number or a \code{n_fields}-length vector.
Defaults to \code{c(3, 0)}.}

\item{wasp_delay}{Delay in days between when the aphids start and the
wasps are added.
This can be a single integer or a \code{n_fields}-length vector.
Defaults to \code{8}.}

\item{wasp_disp_m0}{Proportion of adult wasps from each field that
are added to the dispersal pool when there are no aphids present.
After adding wasps to the pool, they are then evenly distributed
to all fields.
This happens only on days indicated by \code{plant_check_gaps}.
Defaults to \code{0}.}

\item{wasp_disp_m1}{Effect of aphid density on wasp emigration from a patch.
Emigration is \code{wasp_disp_m0 * exp(-wasp_disp_m1 * z)}, where \code{z} is
the total number of living aphids in the patch.
Defaults to \code{0}.}

\item{wasp_field_attract}{Relatively attractiveness of fields to wasps.
This affects the proportion of wasps that immigrate from the dispersal
pool to each field.
It doesn't change the number of wasps that leave fields.
This can be a single numeric or a numeric vector of length \code{n_fields}.
If a single numeric is provided, all fields are equally attractive
to wasps.
If a vector is provided, then the vector is divided by its sum
(to make it sum to 1), then those values are used as the proportion of
wasps from the dispersal pool that immigrate to each field.
Defaults to \code{1}.}

\item{constant_wasps}{Logical for whether to keep adult wasps at the same
density throughout simulations.
This can be a single logical or a \code{n_fields}-length vector.
Defaults to \code{FALSE}.}

\item{mum_smooth}{Proportion of mummies that will NOT take exactly 3 days
to develop. As this value approaches 2/3, it will provide greater
smoothing of wasp numbers through time.
Defaults to \code{0.4}.}

\item{pred_rate}{Daily predation rate on aphids and mummies.
Defaults to \code{0.1} to compensate for losses from plants dying.}

\item{extinct_N}{Threshold below which a line is considered extinct.
Defaults to \code{1}.}

\item{save_every}{Abundances will be stored every \code{save_every} time points.
Defaults to \code{1}.}

\item{perturb}{Information for perturbing populations in the simulations.
It should be a dataframe with 4 columns:
* \code{when}: Integers indicating at what timepoint(s) to do the
perturbations. These can be repeated if you want to perturb
multiple things at the same time.
* \code{where}: What field to do the perturbations in.
* \code{who}: Which population to perturb.
This can be a character vector where values must be the name of an
aphid line, \code{"wasps"}, or \code{"mummies"}.
It can also be an integer vector where, for \code{n} aphid lines,
values \verb{<= n} indicate an aphid line,
values \verb{== n+1} indicate mummies,
and values \verb{== n+2} indicate adult wasps.
Note that perturbing the mummy population also perturbs the
still-living but parasitized aphids, too.
* \code{how}: Numbers \verb{>= 0} that are multiplied by the desired population
to cause the perturbation.}

\item{show_progress}{Boolean for whether to show progress bar.
Defaults to \code{FALSE}.}
}
\description{
Deterministic simulations of multiple fields of aphids and wasps.
}
