% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{sim_fields}
\alias{sim_fields}
\title{Simulate fields of aphids and natural enemies.}
\usage{
sim_fields(
  fields,
  n_reps = 1,
  max_t = 250,
  perturb = NULL,
  save_every = 1,
  sep_adults = FALSE,
  stage_ts_out = FALSE,
  show_progress = FALSE,
  n_threads = 1L
)
}
\arguments{
\item{fields}{An \code{AllFields} object created using the
\code{all_fields} function.}

\item{n_reps}{Number of reps to simulate. Defaults to \code{1}.}

\item{max_t}{How many days to simulate. Defaults to \code{250}.}

\item{perturb}{Information for perturbing populations in the simulations.
It should be a dataframe with 4 columns:
* \code{when}: Integers indicating at what timepoint(s) to do the
perturbations. These can be repeated if you want to perturb
multiple things at the same time.
* \code{where}: What field to do the perturbations in.
* \code{who}: Which population to perturb.
This can be a character vector where values must be the name of an
aphid line, \code{"wasps"}, or \code{"mummies"}.
It can also be an integer vector where, for \code{n} aphid lines,
values \verb{<= n} indicate an aphid line,
values \verb{== n+1} indicate mummies,
and values \verb{== n+2} indicate adult wasps.
Note that perturbing the mummy population also perturbs the
still-living but parasitized aphids, too.
* \code{how}: Numbers \verb{>= 0} that are multiplied by the desired population
to cause the perturbation.}

\item{save_every}{Abundances will be stored every \code{save_every} time points.
Defaults to \code{1}.}

\item{sep_adults}{Single logical for whether to separate adults vs
juvenile aphids in the output.
Useful for counting aphids with wings (adult alates).
Defaults to \code{FALSE}.}

\item{stage_ts_out}{Single logical for whether to output stage-structured
information for all time points.
If \code{TRUE}, the output object will contain this information in the
\code{stage_ts} field, which will be a list of data frames.
Defaults to \code{FALSE}.}

\item{show_progress}{Boolean for whether to show progress bar.
Defaults to \code{FALSE}.}

\item{n_threads}{Number of threads to use if \code{n_reps > 1}
(ignored otherwise).
Defaults to \code{1}.}
}
\value{
A list of dataframes of aphids, wasps.
}
\description{
Simulate fields of aphids and natural enemies.
}
\examples{
wp <- wasp_pop()
sa <- clonal_line("susceptible")
ra <- clonal_line("resistant", resistant = TRUE, repro_apterous = "low")
af <- all_fields(c(sa, ra), wp)
sf <- sim_fields(af)


}
