% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{sim_reps}
\alias{sim_reps}
\title{Simulate multiple reps and simplify output.}
\usage{
sim_reps(n_reps, max_t = 1000, N0 = NULL, R = NULL, A = NULL,
  D_vec = NULL, process_error = NULL, disp_error = TRUE,
  log_zeta_mean = NULL, log_zeta_sd = NULL, mu_time = NULL,
  repl_times = NULL, repl_threshold = 500, extinct_N = 1e-04,
  save_every = max_t\%/\%100, by_patch = FALSE, n_cores = 1,
  show_progress = FALSE, line_names = NULL)
}
\arguments{
\item{n_reps}{Number of reps to simulate.}

\item{max_t}{Max time points to simulate for each rep.}

\item{R}{Growth rates for each line.}

\item{A}{Density dependence for each line.}

\item{process_error}{SD of process error.}

\item{repl_times}{Vector of dates on which to replace plants. Use 0-based indexing!}

\item{n_cores}{Number of cores to use. Defaults to \code{1}.}

\item{show_progress}{Boolean for whether to show progress bar. Defaults to
\code{FALSE}.}

\item{N_0}{Starting abundances for each aphid line on each plant.}

\item{D_binom}{Data frame containing intercept and coefficient estimates for
binomial GLM of \code{<whether dispersal > 0> ~ N_t + N_t^2}.
There should be one set of estimates for each line regardless of whether the model
estimated separate values for each line.}

\item{D_nb}{Data frame containing intercept and overdispersion estimates for
negative binomial GLM of \code{<# dispersed | dispersal occurs> ~ 1}.
There should be one set of estimates for each line regardless of whether the model
estimated separate values for each line.}

\item{plant_mort_0}{Intercept for plant-death-induced aphid mortality, one per line.}

\item{plant_mort_1}{Coefficient for plant-death-induced aphid mortality, one per line.}

\item{plant_death_age_mean}{Mean of distribution of plant-death ages.}

\item{plant_death_age_sd}{SD of distribution of plant-death ages.}

\item{repl_age}{Number of days past death when a plant will be replaced.}
}
\description{
Simulate multiple reps and simplify output.
}
