% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{sim_stochastic}
\alias{sim_stochastic}
\title{Stochastic simulations of multiple fields of aphids and wasps.}
\usage{
sim_stochastic(
  clonal_lines,
  n_fields = 2,
  max_t = 250,
  n_plants = 16,
  K = 1800,
  pred_rate = 0,
  alate_field_disp_p = 0.35,
  aphid_plant_disp_p = 0.1,
  plant_check_gaps = c(3, 4),
  wilted_N = 500,
  aphid_demog_error = FALSE,
  wasp_demog_error = FALSE,
  environ_error = FALSE,
  sigma_x = environ$sigma_x,
  sigma_y = environ$sigma_y,
  rho = environ$rho,
  n_reps = 1,
  n_threads = 1,
  ...
)
}
\arguments{
\item{clonal_lines}{A \code{multiAphid} object containing the aphid-line-specific
info for all the lines in the simulations.
Each object in the \code{multiAphid} is an \code{aphid} object that results from
the \code{clonal_line} function.
Combine them using \code{c(aphid_obj1, aphid_obj2)}.}

\item{n_fields}{The number of fields to simulate.
Wasps operate at the field level, whereas aphids operate at the
plant level.
Both wasps and aphids operate separately across fields but can be
connected via dispersal (see arguments \code{wasp_disp_m0}, \code{wasp_disp_m1},
and \code{alate_field_disp_p}).
Defaults to \code{2}.}

\item{max_t}{How many days to simulate. Defaults to \code{250}.}

\item{n_plants}{The number of plants per field.
This argument is only useful if you want to simulate the process
of plants or groups of plants dying.
Wasps operate at the field level, whereas aphids operate at the
plant level.
"Plant" can indicate a group of plants where the aphids can freely
disperse across them.
Aphids can disperse across plants via the \code{aphid_plant_disp_p} argument.
Defaults to \code{16}.}

\item{K}{Aphid density dependence.
Defaults to \code{1800} because this caused simulations to
approximately match growth of aphid populations on individual fava
bean plants.}

\item{pred_rate}{Daily predation rate on aphids and mummies.
Defaults to \code{0} because we're simulating an experiment with no predators
and, in contrast to the \code{sim_experiments} function, we're
explicitly simulating the process of plants dying and being replaced.}

\item{alate_field_disp_p}{Proportion of alates from each field that
are added to the dispersal pool.
After adding alates to the pool, they are then evenly distributed
to all fields.
This happens only on days indicated by \code{plant_check_gaps}.
The default here is greater than that in \code{sim_experiments} because
these simulations only exchange alates twice per week instead of
every day.
Defaults to \code{0.35}.}

\item{aphid_plant_disp_p}{Proportion of aphids added to the between-plant
dispersal pool every day. Note that this does not depend on
\code{plant_check_gaps} like among-field alate dispersal does.
This can be a single number or a vector with the same length as the
number of aphid lines.
Defaults to \code{0.1}.}

\item{plant_check_gaps}{Gap(s) between when plants are check on.
This is used if you want to see what will happen if you can't check
on things every day in an experiment.
You could use \code{c(3, 4)} for checking twice per week, for example.
Note that this argument determines how often
(1) wasps and alates disperse across fields and
(2) plants are checked for being wilted and potentially replaced.
So using this argument to simulate harvesting fields will also cause
wasp and alate dispersal to not occur daily.
We included this argument because we would have to replicate
among-field dispersal by hand in the experiments we were planning,
and we wanted to be sure that not checking daily wouldn't affect
anything significantly.
It didn't, which is why \code{sim_experiments} has checks every day.
Defaults to \code{c(3, 4)}.}

\item{wilted_N}{Total number of aphids that causes the plant
to become wilted. Values <= 0 cause this to be ignored.
Defaults to \code{500}.}

\item{aphid_demog_error}{Logical for whether to have demographic
stochasticity for aphids. Defaults to \code{FALSE}.}

\item{wasp_demog_error}{Logical for whether to have demographic
stochasticity for wasps. Defaults to \code{FALSE}.}

\item{environ_error}{Logical for whether to have environmental stochasticity.
Defaults to \code{FALSE}.}

\item{sigma_x}{Standard deviation of environmental stochasticity for aphids.
Defaults to the internal object \code{environ$sigma_x},
which is from Meisner et al. (2014).}

\item{sigma_y}{Standard deviation of environmental stochasticity for wasps.
Defaults to the internal object \code{environ$sigma_y},
which is from Meisner et al. (2014).}

\item{rho}{Environmental correlation among instars.
Defaults to \code{environ$rho}.}

\item{n_reps}{Number of reps to simulate. Defaults to \code{1}.}

\item{n_threads}{Number of threads to use if OpenMP is enabled
(ignored otherwise).
Find out whether it's enabled using \code{gameofclones:::using_openmp()}.
Defaults to \code{1}.}
}
\description{
Stochastic simulations of multiple fields of aphids and wasps.
}
\seealso{
\code{\link[=sim_experiments]{sim_experiments()}}
}
