% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stan_estimates}
\alias{stan_estimates}
\title{Estimates of the stan model of population parameters.}
\format{A list with the following items:
\describe{
\item{names}{A length-8 character vector of names for each aphid line.}
\item{R}{A length-8 numeric vector of growth rates for each aphid line.}
\item{A}{A length-8 numeric vector of density dependences for each aphid line.}
\item{process_error}{SD of process error.}
\item{log_zeta_mean}{Mean of log(zeta) distribution.}
\item{log_zeta_sd}{SD of log(zeta) distribution.}
\item{mu_time}{Mean time among all time series. Used to calculate the effect of
plant deterioration through time.}
}}
\usage{
stan_estimates
}
\description{
Estimates of the stan model of population parameters.
}
\keyword{datasets}
