% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{target_type_sims}
\alias{target_type_sims}
\title{Simulate targets of different types}
\usage{
target_type_sims(x_size, y_size, wt_mat, n_samples, fill_all = FALSE)
}
\arguments{
\item{x_size}{Single numeric indicating x dimension of search area.
Bounds will be \code{c(0, x_size)}.}

\item{y_size}{Single numeric indicating y dimension of search area.
Bounds will be \code{c(0, y_size)}.}

\item{wt_mat}{Square numeric matrix indicating how sample weighting on
neighboring locations is affected by a target of each type being
placed in a particular spot.
Item \code{wt_mat[i,j]} indicates the effect of target type \code{i} on
subsequent samplings of target type \code{j}.
Values above 1 cause neighboring locations to be more likely to be
sampled later, while values below 1 cause them to be less likely
sampled.
For locations that have been adjusted using \code{wt_mat} multiple times,
the weights are multiplied by each other
(e.g., \verb{w *= wt_mat[1,3]} at time \code{t}, then
\verb{w *= wt_mat[1,2]} at time \code{t+1}).
All weights start with values of \code{1}.
The matrix should have the same number of rows and columns as the
number of target types.}

\item{n_samples}{Integer vector indicating the number of samples to
produce per target type. The length should equal the number of
target types, and all values should be \verb{>=1}.}

\item{fill_all}{Single logical indicating whether to include a row for
all points in the landscape. If \code{TRUE}, then locations where no targets
exist will contain an empty vector in the \code{type} column.
Defaults to \code{FALSE}.}
}
\value{
A \code{\link[tibble:tbl_df-class]{tibble}} with columns \code{x}, \code{y}, and \code{type}.
}
\description{
Simulate locations of targets of different types along an evenly spaced
grid of integers, where the placement of one target can affect subsequent
placement of other targets.
}
