% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{theta}
\alias{theta}
\title{Priors for use in estimating population parameters.}
\format{A numeric vector of length 12. Each number is a hyperparameter value.
See \code{vignette("choose_priors")} for more information.}
\usage{
theta
}
\description{
Priors for use in estimating population parameters.
}
\keyword{datasets}
