% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wasp_attack}
\alias{wasp_attack}
\title{Wasp attack rate parameters.}
\format{
A list with five items:
\describe{
\item{a}{
parasitoid attack rate
}
\item{k}{
aggregation parameter of the negative binomial distribution
}
\item{h}{
parasitoid attack rate handling time
}
\item{rel_attack}{
A vector of length 5, giving  relative attack rates on the
different instars.
}
\item{attack_surv}{
A numeric vector of length two, of survivals of singly attacked and
multiply attacked resistant aphids.
\emph{Note:} This is not from either paper, but from unpublished code by
Anthony Ives.
}
}
}
\source{
\url{http://doi.wiley.com/10.1890/13-1933.1}

\url{http://www.journals.uchicago.edu/doi/10.1086/303269}
}
\usage{
wasp_attack
}
\description{
Wasp attack rate parameters.
}
\keyword{datasets}
