% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aphid-wasp-classes.R
\name{wasp_pop}
\alias{wasp_pop}
\title{Organize wasp population info.}
\usage{
wasp_pop(
  s_y = populations$s_y,
  sex_ratio = populations$sex_ratio,
  a = wasp_attack$a,
  k = wasp_attack$k,
  h = wasp_attack$h,
  rel_attack = wasp_attack$rel_attack,
  sigma_y = environ$sigma_y,
  mum_smooth = 0.4
)
}
\arguments{
\item{s_y}{Daily survival rate of adult wasps.
Defaults to \code{populations$s_y}, which is from Meisner et al. (2014).}

\item{sex_ratio}{Sex ratio of adult wasps. Defaults to \code{0.5}.}

\item{a}{Parasitoid attack rate. Defaults to the internal object
\code{wasp_attack$a}, which is from Meisner et al. (2014).}

\item{k}{Aggregation parameter of the negative binomial distribution.
Defaults to the internal object \code{wasp_attack$k},
which is from Meisner et al. (2014).}

\item{h}{Parasitoid handling time. Defaults to the internal object
\code{wasp_attack$h}, which is from Meisner et al. (2014).}

\item{rel_attack}{Relative parasitoid attack rate among instars.
Defaults to \code{wasp_attack$rel_attack}, which is from
Meisner et al. (2014).}

\item{sigma_y}{Standard deviation of environmental stochasticity for wasps.
This argument has no effect if \code{environ_error = FALSE}.
Defaults to the internal object \code{environ$sigma_y},
which is from Meisner et al. (2014).}

\item{mum_smooth}{Proportion of mummies that will NOT take exactly 3 days
to develop. As this value approaches 2/3, it will provide greater
smoothing of wasp numbers through time.
Defaults to \code{0.4}.}
}
\value{
List containing an external pointer of a C++ object that can be
passed onto \code{all_fields}.
}
\description{
Organize wasp population info.
}
