% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{clonal_line}
\alias{clonal_line}
\title{Organize clonal line information}
\usage{
clonal_line(
  name,
  density_0 = NULL,
  resistant = FALSE,
  surv_juv_apterous = NULL,
  surv_adult_apterous = NULL,
  repro_apterous = NULL,
  surv_juv_alates = "low",
  surv_adult_alates = "low",
  repro_alates = "low",
  surv_juv_paras = "low",
  surv_adult_paras = "low",
  repro_paras = "low",
  temp = "low"
)
}
\arguments{
\item{name}{String for clonal line name.}

\item{density_0}{A 5 x 2 matrix with the rows indicating aphid instar,
and column indicating apterous vs alate.
Defaults to \code{NULL}, which results in four 4th instars.}

\item{resistant}{Logical or length-2 vector of survivals of
singly attacked and multiply attacked aphids.
If a logical, \code{FALSE} is equivalent to \code{c(0,0)} and results in no
resistance.
\code{TRUE} results in the resistance values for a resistance line
from unpublished work by Anthony Ives.
Defaults to \code{FALSE}.}

\item{surv_juv_apterous}{A single number for the juvenile survival rate for
apterous aphids. Defaults to \code{NULL}, which results in estimates
from a medium-reproduction line.}

\item{surv_adult_apterous}{A vector of adult survival probabilities for
apterous aphids. Defaults to \code{NULL}, which results in estimates
from a medium-reproduction line.}

\item{repro_apterous}{A vector of fecundities for for apterous aphids.
Defaults to \code{NULL}, which results in estimates from a
medium-reproduction line.}

\item{surv_juv_alates}{A single number for the juvenile survival rate for
alates aphids. Defaults to \code{"low"}, which results in estimates
from a low-reproduction line.}

\item{surv_adult_alates}{A vector of adult survival probabilities for
alates aphids. Defaults to \code{"low"}, which results in estimates
from a low-reproduction line.}

\item{repro_alates}{A vector of fecundities for for alates aphids.
Defaults to \code{"low"}, which results in estimates from a low-reproduction
line.}

\item{surv_juv_paras}{A single number for the juvenile survival rate for
paras aphids. Defaults to \code{"low"}, which results in estimates
from a low-reproduction line.}

\item{surv_adult_paras}{A vector of adult survival probabilities for
paras aphids. Defaults to \code{"low"}, which results in estimates
from a low-reproduction line.}

\item{repro_paras}{A vector of fecundities for for paras aphids.
Defaults to \code{"low"}, which results in estimates from a low-reproduction
line.}

\item{temp}{Single string specifying \code{"low"} (20º C) or \code{"high"} (27º C)
temperature. Defaults to \code{"low"}.}
}
\value{
A list with the necessary info to pass onto sim_clonewars.
}
\description{
Organize clonal line information
}
