% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_models.R
\name{fit_lines}
\alias{fit_lines}
\title{Fit multiple time series for multiple aphid lines.}
\usage{
fit_lines(data, X, L, ...)
}
\arguments{
\item{data}{An optional data frame, list, or environment that contains `X`.
By default, variables are taken from the environment from which
the function was called.}

\item{X}{Name of the matrix in `data` that contains the log-transformed counts
through time (1 time series per column).
This option is not required if the proper object inside `data` is literally
named `X` (as would be the case if you used `line_data()`).
Each column should contain `NA`s at the end (ONLY the end) if
it wasn't observed as many times as was the time series in the matrix with
the most observations.
\emph{Other than at the end, missing values are not yet supported}.}

\item{L}{Vector (of the same length as number of columns in `X`) containing
the line number for each time-series column in `X`.
This option is not required if the proper object inside `data` is literally
named `L` (as would be the case if you used `line_data()`).}

\item{...}{Arguments passed to `rstan::sampling` (e.g., iter, chains).}
}
\value{
A `stanfit` object containing the model fit.
}
\description{
Fit multiple time series for multiple aphid lines.
}
