% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_models.R
\name{fit_lines}
\alias{fit_lines}
\title{Fit multiple time series for multiple aphid lines.}
\usage{
fit_lines(data_df, line, rep, date, X, ...)
}
\arguments{
\item{data_df}{A data frame that contains columns for \code{line}, \code{rep}, and \code{X}.}

\item{line}{An optional string specifying the name for the column in \code{data_df} that
contains which line each observation belongs to.
Defaults to \code{"line"}.}

\item{rep}{An optional string specifying the name for the column in \code{data_df} that
contains which rep each observation belongs to.
Defaults to \code{"rep"}.}

\item{date}{An optional string specifying the name for the column in \code{data_df} that
contains the date for each observation.
Defaults to \code{"date"}.}

\item{X}{An optional string specifying the name for the column in \code{data_df} that
contains the \code{X} value for each observation (\code{X = log(N)}).
Defaults to \code{"X"}.}

\item{...}{Arguments passed to \code{rstan::sampling} (e.g., iter, chains).}
}
\value{
A \code{stanfit} object containing the model fit.
}
\description{
Fit multiple time series for multiple aphid lines.
}
