% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_models.R
\name{fit_one_line}
\alias{fit_one_line}
\title{Fit multiple time series for a single aphid line.}
\usage{
fit_one_line(X, data = sys.frame(sys.parent()), ...)
}
\arguments{
\item{X}{Name of the matrix in `data` that contains the log-transformed counts
through time (1 time series per column).
Each column should contain `NA`s at the end (ONLY the end) if
it wasn't observed as many times as was the time series in the matrix with
the most observations.
\emph{Other than at the end, missing values are not yet supported}.}

\item{data}{An optional data frame, list, or environment that contains `X`.
By default, variables are taken from the environment from which
the function was called.}

\item{...}{Arguments passed to `rstan::sampling` (e.g., iter, chains).}
}
\value{
A `stanfit` object containing the model fit.
}
\description{
Fit multiple time series for a single aphid line.
}
\examples{

R0 <- 0.29
alpha <- 0.002
process <- 0.1
n_reps <- 5
nobs_ts <- 11:15

X <- sim_lines(R0, alpha, n_reps, nobs_ts, process)

# (You would normally not change iter to 100)
fit <- fit_one_line(X, iter = 100)

}
