% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gameofclones-package.R
\docType{package}
\name{gameofclones-package}
\alias{gameofclones-package}
\alias{gameofclones}
\title{The 'gameofclones' package.}
\description{
Analyses and simulations related to aphid--parasitoid
eco-evo dynamics experiments.
}
