% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manip.R
\name{load_data}
\alias{load_data}
\title{Load aphid growth data.}
\usage{
load_data(file, noNA = TRUE, filter_pars = list(begin = 0.5, end = 0.9))
}
\arguments{
\item{file}{A filename to read from. If left empty, this will read from the
default path on your DropBox folder.}

\item{noNA}{Boolean for whether time series with NAs should be included.
Defaults to `TRUE`.}

\item{filter_pars}{A list with the names `"begin"` and `"end"`, containing single
numbers with threshold for filtering the beginning and ending of time series,
respectively. Set to `NULL` to avoid filtering entirely.
Defaults to `list(begin = 0.5, end = 1.0)`.}
}
\value{
A data frame containing the aphid population-growth data.
}
\description{
Load aphid growth data.
}
\examples{

growth <- load_data()

}
