% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{restart_experiment}
\alias{restart_experiment}
\title{Restart experimental simulations.}
\usage{
restart_experiment(
  sims_obj,
  new_starts = NULL,
  max_t = 250,
  save_every = 1,
  alate_field_disp_p = NULL,
  K = NULL,
  alate_b0 = NULL,
  alate_b1 = NULL,
  K_y_mult = NULL,
  s_y = NULL,
  a = NULL,
  k = NULL,
  h = NULL,
  wasp_disp_p = NULL,
  mum_smooth = NULL,
  pred_rate = NULL,
  plant_check_gaps = NULL,
  max_plant_age = NULL,
  clear_surv = NULL,
  show_progress = NULL
)
}
\arguments{
\item{sims_obj}{A \code{cloneSims} object output from \code{sim_experiments}.}

\item{new_starts}{A dataframe or list of dataframes indicating the
new starting abundances for all populations (wasps, mummies,
all aphid lines) and stages.
It should be the exact same format as what's in \code{sims_obj$all_info}.
(But don't change \code{sims_obj$all_info} to make this be true!)}

\item{max_t}{How many days to simulate. Defaults to \code{250}.}

\item{save_every}{Abundances will be stored every \code{save_every} time points.
Defaults to \code{1}.}

\item{alate_field_disp_p}{Proportion of alates from each field that
are added to the dispersal pool.
After adding alates to the pool, they are then evenly distributed
to all fields.
This happens only on days indicated by \code{plant_check_gaps}.
Defaults to \code{0.1}.}

\item{K}{Aphid density dependence.
Defaults to \code{12.5e3} because this caused simulations to
approximately match experiments.}

\item{alate_b0}{The proportion of offspring from apterous aphids is
\code{inv_logit(alate_b0 + alate_b1 * N)} where \code{N} is the total number of
aphids on that plant.
Defaults to \code{-5} because we're making this weakly density dependent.}

\item{alate_b1}{The proportion of offspring from apterous aphids is
\code{inv_logit(alate_b0 + alate_b1 * N)} where \code{N} is the total number of
aphids on that plant.
Defaults to \code{K * 1.76e-07}, which makes alate production only mildly
density dependent.}

\item{K_y_mult}{The number multiplied by \code{K} to get density dependence for
parasitized aphids.
Defaults to \code{1 / 1.57}, which is from Meisner et al. (2014).}

\item{s_y}{Daily survival rate of adult wasps.
Defaults to \code{populations$s_y}, which is from Meisner et al. (2014).}

\item{a}{Parasitoid attack rate. Defaults to the internal object
\code{wasp_attack$a}, which is from Meisner et al. (2014).}

\item{k}{Aggregation parameter of the negative binomial distribution.
Defaults to the internal object \code{wasp_attack$k},
which is from Meisner et al. (2014).}

\item{h}{Parasitoid handling time. Defaults to the internal object
\code{wasp_attack$h}, which is from Meisner et al. (2014).}

\item{wasp_disp_p}{Proportion of adult wasps from each field that
are added to the dispersal pool.
After adding wasps to the pool, they are then evenly distributed
to all fields.
This happens only on days indicated by \code{plant_check_gaps}.
Defaults to \code{0}.}

\item{mum_smooth}{Proportion of mummies that will NOT take exactly 3 days
to develop. As this value approaches 2/3, it will provide greater
smoothing of wasp numbers through time.
Defaults to \code{0.4}.}

\item{pred_rate}{Daily predation rate on aphids and mummies.
Defaults to \code{0.1} to compensate for losses from plants dying.}

\item{plant_check_gaps}{Gap(s) between when plants are check on.
This is used if you want to see what will happen if you can't check
on things every day in an experiment.
You could use \code{c(3, 4)} for checking twice per week, for example.
Note that this argument determines how often
(1) wasps and alates disperse across fields and
(2) plants are check for exceeding \code{max_plant_age} or \code{max_N}.
So using this argument to simulate harvesting fields will also cause
wasp and alate dispersal to not occur daily.
Defaults to \code{1}.}

\item{max_plant_age}{Age at which plants are cleared.
This can be useful to simulate harvesting fields.
A value of \code{0} turns this off completely, but it also requires that
\code{max_N > 0} because the internal code wants some threshold to look for.
Defaults to \code{1e9} which effectively turns this off.}

\item{clear_surv}{Survival of aphids and mummies when a harvest occurs.
Defaults to \code{0}.}

\item{show_progress}{Boolean for whether to show progress bar.
Defaults to \code{FALSE}.}
}
\description{
Note that defaults for all this function's arguments except for
\code{sims_obj}, \code{new_starts}, \code{max_t}, and \code{save_every} are \code{NULL},
which results in them being the same as for the original simulations.
}
