% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_test_data.R
\name{sim_lines}
\alias{sim_lines}
\title{Simulate log-transformed counts of aphid populations for one or more lines.}
\usage{
sim_lines(R0, alpha, n_reps, nobs_ts, sigma_process = 0.1)
}
\arguments{
\item{R0}{Vector of population growth rates for each line.}

\item{alpha}{Vector of density dependencies for each line.}

\item{n_reps}{Number of reps per aphid line.}

\item{nobs_ts}{Number of observations per time series.}

\item{sigma_process}{Process error. Defaults to `0.1`.}
}
\description{
Simulate log-transformed counts of aphid populations for one or more lines.
}
