% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{sim_stochastic}
\alias{sim_stochastic}
\title{Stochastic simulations of multiple fields of aphids and wasps.}
\usage{
sim_stochastic(
  clonal_lines,
  n_fields = 2,
  max_t = 250,
  wasp_badger_n = 0,
  aphid_demog_error = TRUE,
  wasp_demog_error = TRUE,
  environ_error = FALSE,
  sigma_x = environ$sigma_x,
  sigma_y = environ$sigma_y,
  rho = environ$rho,
  n_reps = 1,
  n_threads = 1,
  ...
)
}
\arguments{
\item{clonal_lines}{A \code{multiAphid} object containing the aphid-line-specific
info for all the lines in the simulations.
Each object in the \code{multiAphid} is an \code{aphid} object that results from
the \code{clonal_line} function.
Combine them using \code{c(aphid_obj1, aphid_obj2)}.}

\item{n_fields}{The number of fields to simulate.
Wasps operate at the field level, whereas aphids operate at the
plant level.
Both wasps and aphids operate separately across fields but can be
connected via dispersal (see arguments \code{wasp_disp_m0}, \code{wasp_disp_m1},
and \code{alate_field_disp_p}).
Defaults to \code{2}.}

\item{max_t}{How many days to simulate. Defaults to \code{250}.}

\item{wasp_badger_n}{The number of aphids each adult wasp badgers and kills
per time step. Defaults to \code{0}.}

\item{aphid_demog_error}{Logical for whether to have demographic
stochasticity for aphids. Defaults to \code{FALSE}.}

\item{wasp_demog_error}{Logical for whether to have demographic
stochasticity for wasps. Defaults to \code{FALSE}.}

\item{environ_error}{Logical for whether to have environmental stochasticity.
This argument applies to both aphids and wasps, but if you want to
have just one species experience environmental stochasticity,
then you can set their associated \verb{sigma_*} argument to zero.
Defaults to \code{FALSE}.}

\item{sigma_x}{Standard deviation of environmental stochasticity for aphids.
This argument has no effect if \code{environ_error = FALSE}.
Defaults to the internal object \code{environ$sigma_x},
which is from Meisner et al. (2014).}

\item{sigma_y}{Standard deviation of environmental stochasticity for wasps.
This argument has no effect if \code{environ_error = FALSE}.
Defaults to the internal object \code{environ$sigma_y},
which is from Meisner et al. (2014).}

\item{rho}{Environmental correlation among instars.
Defaults to \code{environ$rho}.}

\item{n_reps}{Number of reps to simulate. Defaults to \code{1}.}

\item{n_threads}{Number of threads to use if OpenMP is enabled
(ignored otherwise).
Find out whether it's enabled using \code{pseudogameofclones:::using_openmp()}.
Defaults to \code{1}.}
}
\description{
Stochastic simulations of multiple fields of aphids and wasps.
}
\seealso{
\code{\link[=sim_experiments]{sim_experiments()}}
}
