% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{target_type_sims}
\alias{target_type_sims}
\title{Simulate targets of different types}
\usage{
target_type_sims(
  x_size,
  y_size,
  wt_mat,
  n_samples,
  allow_overlap = TRUE,
  fill_all = TRUE,
  n_lands = 1L,
  show_progress = FALSE,
  n_threads = 1L
)
}
\arguments{
\item{x_size}{Single integer indicating x dimension of search area.
Locations will be drawn from \code{1} to \code{x_size-1}.
See description above for why.}

\item{y_size}{Single integer indicating y dimension of search area.
Locations will be drawn from \code{1} to \code{y_size-1}.
See description above for why.}

\item{wt_mat}{Square numeric matrix indicating how sample weighting on
neighboring locations is affected by a target of each type being
placed in a particular spot.
Item \code{wt_mat[i,j]} indicates the effect of target type \code{i} on
subsequent samplings of target type \code{j}.
Values above 1 cause neighboring locations to be more likely to be
sampled later, while values below 1 cause them to be less likely
sampled.
For locations that have been adjusted using \code{wt_mat} multiple times,
the weights are multiplied by each other
(e.g., \verb{w *= wt_mat[1,3]} at time \code{t}, then
\verb{w *= wt_mat[1,2]} at time \code{t+1}).
All weights start with values of \code{1}.
The matrix should have the same number of rows and columns as the
number of target types.}

\item{n_samples}{Integer vector indicating the number of samples to
produce per target type. The length should equal the number of
target types, and all values should be \verb{>=1}.}

\item{allow_overlap}{Single logical indicating whether to allow
targets to be more than one type.
An example of a target being multiple types would be a plant that
hosts an epiphytic bacteria and is also infected with a virus.
Defaults to \code{TRUE}.}

\item{fill_all}{Single logical indicating whether to have a target at all
points in the landscape. If \code{TRUE}, then locations where no targets
were simulated will be assigned to a target type \code{length(n_samples)+1}.
Defaults to \code{TRUE}.}

\item{n_lands}{Single integer indicating the number of independent
landscapes to simulate. Each landscape will have a number of samples
per type according to \code{n_samples}.
Must be \verb{> 0} and \verb{< 1e6}.
Defaults to \code{1}.}

\item{show_progress}{Single logical for whether to show progress bar.
Defaults to \code{FALSE}.}

\item{n_threads}{Single integer for the number of threads to use.
Ignored if \code{n_lands == 1}.
Defaults to \code{1L}.}
}
\value{
A list of length \code{n_lands}, where each item is a
\code{\link[tibble:tbl_df-class]{tibble}} with columns \code{x}, \code{y}, and \code{type}.
}
\description{
Simulate locations of targets of different types along an evenly spaced
grid of integers, where the placement of one target can affect subsequent
placement of other targets.
Target locations are drawn from all combinations of \code{1} to
\code{x_size-1} and \code{1} to \code{y_size-1}.
I'm subtracting 1 because the landscape in \code{searcher_sims} is defined by
the bounds \code{c(0, x_size)} and \code{c(0, y_size)}, so this subtraction
keeps a gap of 1 between the most extreme locations and the bounds
of the landscape.
This prevents targets from being located on the bounds.
}
